swap button implemnted

+++ app/src/main/java/com/ahmedelgendy/banquemisrtask/convert/data/model/convert/Query.kt
@@ -1,7 +1,7 @@
 package com.ahmedelgendy.banquemisrtask.convert.data.model.convert
 
 data class Query(
-    val amount: Int? = null,
+    val amount: Double? = null,
     val from: String? = null,
     val to: String? = null
 )
